HNG-Stage-2/RESTApi/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
# drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
Markdown==3.4.4
mysqlclient==2.2.0
packaging==23.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
